services/wallet-service/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
SQLAlchemy==2.0.9
psycopg2-binary==2.9.6
python-dotenv==1.0.0
web3==6.5.0
pyjwt==2.6.0
pika==1.3.1
requests==2.31.0
gunicorn==20.1.0
prometheus-flask-exporter==0.22.3 